Went throught the full installation process and cloned into separate repo, made edits to migrations so no failures occur. original migrations in migrations.bak

+++ app/database/migrations/2014_10_24_232907_update_events_table.php
@@ -12,7 +12,10 @@ class UpdateEventsTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::drop('events');
+		if (Schema::hasTable('events'))
+		{
+			Schema::drop('events');
+		}
 
 		Schema::create('events', function(Blueprint $table)
 		{
